Will not download libflashplayer.so if it exists

+++ media/spotify-apps-installer.sh
@@ -3,7 +3,6 @@
 WRAPPER_BIN="/usr/bin/spotify";
 SPOTIFY_BIN="/usr/bin/spotify.bin";
 SPOTIFY_DEB="http://download.spotify.com/preview/spotify-client_0.8.2.572.geb65f9a.433-1_amd64.deb";
-SPOTIFY_CACHE="$HOME/.cache/spotify";
 SPOTIFY_PLUGIN_PATH="/opt/spotify/plugin";
 FLASH_URL="http://download.macromedia.com/pub/labs/flashplayer10/flashplayer10_2_p3_64bit_linux_111710.tar.gz";
 TMP="/tmp/spotify-installer.tmp";
@@ -21,11 +20,15 @@ function download() {
 }
 
 if [ "$USER" == "root" ]; then
-    mkdir -p $SPOTIFY_PLUGIN_PATH;
-    mv libflashplayer.so $SPOTIFY_PLUGIN_PATH;
-    apt-get install libqt4-webkit;
-    dpkg -i $TMP;
-    mv $WRAPPER_BIN $SPOTIFY_BIN;
+    if [ ! -e "$SPOTIFY_PLUGIN_PATH/libflashplayer.so" ]; then
+        mkdir -p $SPOTIFY_PLUGIN_PATH;
+        mv libflashplayer.so $SPOTIFY_PLUGIN_PATH;
+    fi
+    if [ -z "$SKIP_DOWNLOAD" ]; then
+        apt-get install libqt4-webkit;
+        dpkg -i $TMP;
+        mv $WRAPPER_BIN $SPOTIFY_BIN;
+    fi
     echo Generating $WRAPPER_BIN ...;
     (cat <<CODE
 #!/bin/sh
@@ -36,11 +39,12 @@ CODE
     chmod +x $WRAPPER_BIN;
 
 else
-    mkdir -p $SPOTIFY_CACHE;
-    zenity --question --text "Do you want to install spotify with app support?" --title "Install spotify?" || exit $1;
-    download $FLASH_URL;
-    tar xfzv $TMP;
-    download $SPOTIFY_DEB;
+    if [ ! -e "$SPOTIFY_PLUGIN_PATH/libflashplayer.so" ]; then
+        zenity --question --text "Do you want to install spotify with app support?" --title "Install spotify?" || exit $1;
+        download $FLASH_URL;
+        tar xfzv $TMP;
+    fi
+    [ -z "$SKIP_DOWNLOAD" ] && download $SPOTIFY_DEB;
     gksudo -k $0;
     zenity --info --text "Spotify was installed. It can be started from your application menu";
 fi
